# api/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
      "version": "1.0.0",
      "title": "Map-it API",
      "description": "Roadmap tool",
      "license": {
          "name": "MIT",
          "url": "https://opensource.org/licenses/MIT"
      }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    }
  ],
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/account/login": {
      "post": {
        "tags": ["account"],
        "summary": "Logs in to user account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succesfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/UserInfo"
                }
              }
            }
          },
          "400": { "$ref": "#/definitions/ErrorResponse" }
        }
      }
    },
    "/account/verify": {
      "get": {
        "tags": ["account"],
        "summary": "Verify user JWT token",
        "parameters": [{
          "name": "Authorization",
          "in": "header",
          "schema": { "type": "string" },
          "required": true,
          "example": "Bearer PASTE_TOKEN_HERE"
        }],
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string" },
                    "token": {
                      "type": "string",
                      "description": "JWT token created for user"
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/definitions/ErrorResponse" }
        }
      }
    },
    "/account/refresh": {
      "post": {
        "tags": ["account"],
        "summary": "Renew users JWT token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succesfully got a new JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/UserInfo"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token is incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          },
          "400": { "$ref": "#/definitions/ErrorResponse" }
        }
      }
    },
    "/account/logout": {
      "post": {
        "tags": ["account"],
        "summary": "Invalidate users refresh token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succesfully invalidated refresh token"
          }
        }
      }
    },
    "/account/register": {
      "post": {
        "tags": ["account"],
        "summary": "Create new user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succesfully created a new user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          },
          "400": { "$ref": "#/definitions/ErrorResponse" }
        }
      }
    },
...
  },
  "definitions": {
...
    "Error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Error message"
        },
        "data": {
          "type": "object",
          "nullable": true,
          "description": "Additional error data (optional)"
        }
      }
    },
    "ErrorResponse": {
      "description": "Provided data is invalid",
      "content": {
        "application/json": {
          "schema": {
            "$ref": "#/definitions/Error"
          }
        }
      }
    },
...
    "UserInfo": {
      "type": "object",
      "properties": {
        "email": { "type": "string" },
        "token": {
          "type": "string",
          "description": "JWT token created for user"
        },
        "refreshToken": {
          "type": "string",
          "description": "Refresh token used to renew the JWT token"
        },
        "expiresAt": {
          "type": "string",
          "format": "date-time",
          "description": "Expiration date-time of the JWT token"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": { "type": "integer" },
        "email": { "type": "string" },
        "password": { "type": "string" },
        "name": { "type": "string" }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "scheme": "bearer",
        "name": "Authorization",
        "bearerFormat": "JWT",
        "in": "header"
      }
    }
  }
}